refactor: Piece 패키지 이동
