+++ src/main/java/chess/domain/board/Square.java
@@ -1,7 +1,7 @@
 package chess.domain.board;
 
 import chess.domain.piece.Empty;
-import chess.domain.piece.normal.Piece;
+import chess.domain.piece.Piece;
 import chess.domain.piece.property.Color;
 
 import java.util.Map;
