+++ src/main/java/chess/domain/piece/normal/Normal.java
@@ -1,6 +1,7 @@
 package chess.domain.piece.normal;
 
 import chess.domain.board.Position;
+import chess.domain.piece.Piece;
 import chess.domain.piece.property.Color;
 
 import java.util.HashMap;
