+++ src/main/java/chess/domain/piece/pawn/Pawn.java
@@ -1,7 +1,7 @@
 package chess.domain.piece.pawn;
 
 import chess.domain.board.Position;
-import chess.domain.piece.normal.Piece;
+import chess.domain.piece.Piece;
 import chess.domain.piece.property.Color;
 import chess.domain.piece.property.Kind;
