+++ src/main/java/chess/view/KindMapper.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.domain.piece.normal.Piece;
+import chess.domain.piece.Piece;
 import chess.domain.piece.property.Color;
 import chess.domain.piece.property.Kind;
